pyproject.toml
[tool.poetry]
name = "movie-recommendation"
version = "0.1.0"
description = ""
authors = ["Anne-Lene Sax <Anne-Lene.Sax@t-systems.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
numpy = "^2.1.2"
scikit-learn = "^1.5.2"
logging = "^0.4.9.6"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
